pyproject.toml
[project]
name = "apollonft-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "google-adk>=1.0",
  "google-genai>=1.0",
  "qdrant-client>=1.9",
  "groq>=0.11",
  "pydantic-settings>=2.3",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "websockets>=12",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
